Refactor skill URL install maintainability
