+++ crates/ironclaw_host_runtime/Cargo.toml
@@ -51,6 +51,7 @@ jsonschema = { version = "0.45", default-features = false }
 libc = "0.2"
 rust_decimal = { version = "1", features = ["serde", "serde-with-str"] }
 secrecy = "0.10"
+serde = { version = "1", features = ["derive"] }
 serde_json = "1"
 sha2 = "0.10"
 thiserror = "2"
