+++ crates/ironclaw_host_runtime/src/first_party_tools/skill_url_install.rs
@@ -1,16 +1,15 @@
-use std::{
-    collections::{HashSet, VecDeque},
-    io::Read,
-    path::{Component, Path, PathBuf},
-};
+use std::path::PathBuf;
 
 use ironclaw_host_api::{
     NetworkMethod, NetworkPolicy, ResourceUsage, RuntimeDispatchErrorKind, RuntimeHttpEgressError,
     RuntimeHttpEgressReasonCode, RuntimeHttpEgressRequest, RuntimeKind,
 };
-use serde_json::Value;
 
 use crate::{FirstPartyCapabilityError, FirstPartyCapabilityRequest};
+
+mod bundle;
+mod github;
+mod zip_bundle;
 
 const SKILL_URL_RESPONSE_BODY_LIMIT_BYTES: u64 = 10 * 1024 * 1024;
 const SKILL_URL_FETCH_TIMEOUT_MS: u32 = 10_000;
@@ -33,35 +32,6 @@ pub(super) struct SkillUrlPayloadFile {
 }
 
 #[derive(Debug, Clone, PartialEq, Eq)]
-struct ZipSkillBundle {
-    skill_md: String,
-    files: Vec<SkillUrlPayloadFile>,
-    bundle_subdir: Option<String>,
-}
-
-#[derive(Debug, Clone, PartialEq, Eq)]
-struct GitHubRepoRef {
-    owner: String,
-    repo: String,
-    branch: String,
-    subdir: Option<String>,
-}
-
-#[derive(Debug, Clone, PartialEq, Eq)]
-struct GitHubRepoRequest {
-    owner: String,
-    repo: String,
-    tree_segments: Option<Vec<String>>,
-}
-
-#[derive(Debug, Clone, PartialEq, Eq)]
-struct GitHubBlobRequest {
-    owner: String,
-    repo: String,
-    blob_segments: Vec<String>,
-}
-
-#[derive(Debug, Clone, PartialEq, Eq)]
 struct FetchedBytes {
     status: u16,
     body: Vec<u8>,
@@ -73,7 +43,26 @@ pub(super) async fn fetch_skill_url_payload(
     usage: &mut ResourceUsage,
 ) -> Result<SkillUrlPayload, FirstPartyCapabilityError> {
     let parsed = validate_skill_url(url)?;
-    fetch_skill_payload_from_url(request, &parsed, usage).await
+    if let Some(payload) = github::fetch_payload_if_supported(request, &parsed, usage).await? {
+        return Ok(payload);
+    }
+
+    let bytes = fetch_url_bytes(request, &parsed, usage).await?;
+    if bytes.starts_with(b"PK\x03\x04") {
+        let bundle = zip_bundle::extract_skill_bundle_blocking(bytes, None).await?;
+        return Ok(SkillUrlPayload {
+            content: bundle.skill_md,
+            files: bundle.files,
+        });
+    }
+
+    Ok(SkillUrlPayload {
+        content: String::from_utf8(bytes).map_err(|_| {
+            FirstPartyCapabilityError::new(RuntimeDispatchErrorKind::OperationFailed)
+                .with_usage(usage.clone())
+        })?,
+        files: Vec::new(),
+    })
 }
 
 async fn fetch_url_bytes(
@@ -158,741 +147,8 @@ fn validate_skill_url(url: &str) -> Result<url::Url, FirstPartyCapabilityError> 
     Ok(parsed)
 }
 
-async fn fetch_skill_payload_from_url(
-    request: &FirstPartyCapabilityRequest,
-    parsed: &url::Url,
-    usage: &mut ResourceUsage,
-) -> Result<SkillUrlPayload, FirstPartyCapabilityError> {
-    if let Some(blob) = parse_github_blob_ref(parsed) {
-        let raw_url = resolve_github_blob_download_url(request, blob, usage).await?;
-        let bytes = fetch_url_bytes(request, &raw_url, usage).await?;
-        return Ok(SkillUrlPayload {
-            content: String::from_utf8(bytes).map_err(|_| {
-                FirstPartyCapabilityError::new(RuntimeDispatchErrorKind::OperationFailed)
-                    .with_usage(usage.clone())
-            })?,
-            files: Vec::new(),
-        });
-    }
-
-    if let Some(repo) = parse_github_repo_ref(parsed) {
-        return fetch_github_repo_payload(request, parsed.as_str(), repo, usage).await;
-    }
-
-    let bytes = fetch_url_bytes(request, parsed, usage).await?;
-    if bytes.starts_with(b"PK\x03\x04") {
-        let bundle = extract_skill_bundle_from_zip_blocking(bytes, None).await?;
-        return Ok(SkillUrlPayload {
-            content: bundle.skill_md,
-            files: bundle.files,
-        });
-    }
-
-    Ok(SkillUrlPayload {
-        content: String::from_utf8(bytes).map_err(|_| {
-            FirstPartyCapabilityError::new(RuntimeDispatchErrorKind::OperationFailed)
-                .with_usage(usage.clone())
-        })?,
-        files: Vec::new(),
-    })
-}
-
-fn parse_github_blob_ref(parsed: &url::Url) -> Option<GitHubBlobRequest> {
-    if parsed.host_str()? != "github.com" {
-        return None;
-    }
-    let parts = parsed
-        .path_segments()?
-        .filter(|segment| !segment.is_empty())
-        .collect::<Vec<_>>();
-    if parts.len() < 5 || parts[2] != "blob" {
-        return None;
-    }
-    let repo = parts[1].trim_end_matches(".git").to_string();
-    if repo.is_empty() {
-        return None;
-    }
-    Some(GitHubBlobRequest {
-        owner: parts[0].to_string(),
-        repo,
-        blob_segments: parts[3..]
-            .iter()
-            .map(|segment| (*segment).to_string())
-            .collect(),
-    })
-}
-
-fn parse_github_repo_ref(parsed: &url::Url) -> Option<GitHubRepoRequest> {
-    if parsed.host_str()? != "github.com" {
-        return None;
-    }
-    let parts = parsed
-        .path_segments()?
-        .filter(|segment| !segment.is_empty())
-        .collect::<Vec<_>>();
-    if parts.len() < 2 {
-        return None;
-    }
-    let owner = parts[0].to_string();
-    let repo = parts[1].trim_end_matches(".git").to_string();
-    if repo.is_empty() {
-        return None;
-    }
-    if parts.len() == 2 {
-        return Some(GitHubRepoRequest {
-            owner,
-            repo,
-            tree_segments: None,
-        });
-    }
-    if parts.len() >= 4 && parts[2] == "tree" {
-        return Some(GitHubRepoRequest {
-            owner,
-            repo,
-            tree_segments: Some(
-                parts[3..]
-                    .iter()
-                    .map(|segment| (*segment).to_string())
-                    .collect(),
-            ),
-        });
-    }
-    None
-}
-
-fn is_safe_github_component(component: &str) -> bool {
-    !component.is_empty()
-        && component
-            .bytes()
-            .all(|b| b.is_ascii_alphanumeric() || matches!(b, b'.' | b'_' | b'-'))
-}
-
-fn validate_github_repo_components(
-    owner: &str,
-    repo: &str,
-) -> Result<(), FirstPartyCapabilityError> {
-    if is_safe_github_component(owner) && is_safe_github_component(repo) {
-        Ok(())
-    } else {
-        Err(FirstPartyCapabilityError::new(
-            RuntimeDispatchErrorKind::InputEncode,
-        ))
-    }
-}
-
 fn validate_derived_fetch_url(url: &str) -> Result<url::Url, FirstPartyCapabilityError> {
     validate_skill_url(url)
-}
-
-fn build_github_api_base_url(
-    owner: &str,
-    repo: &str,
-) -> Result<url::Url, FirstPartyCapabilityError> {
-    validate_github_repo_components(owner, repo)?;
-    validate_derived_fetch_url(&format!("https://api.github.com/repos/{owner}/{repo}"))
-}
-
-fn build_github_contents_url(
-    owner: &str,
-    repo: &str,
-    path: Option<&str>,
-    git_ref: &str,
-) -> Result<url::Url, FirstPartyCapabilityError> {
-    let mut url = build_github_api_base_url(owner, repo)?;
-    {
-        let mut segments = url.path_segments_mut().map_err(|_| {
-            FirstPartyCapabilityError::new(RuntimeDispatchErrorKind::OperationFailed)
-        })?;
-        segments.push("contents");
-        if let Some(path) = path {
-            for segment in path.split('/').filter(|segment| !segment.is_empty()) {
-                segments.push(segment);
-            }
-        }
-    }
-    url.query_pairs_mut().append_pair("ref", git_ref);
-    Ok(url)
-}
-
-fn build_github_matching_refs_url(
-    owner: &str,
-    repo: &str,
-    namespace: &str,
-    prefix: &str,
-) -> Result<url::Url, FirstPartyCapabilityError> {
-    if !matches!(namespace, "heads" | "tags") || !is_safe_github_component(prefix) {
-        return Err(FirstPartyCapabilityError::new(
-            RuntimeDispatchErrorKind::InputEncode,
-        ));
-    }
-    let mut url = build_github_api_base_url(owner, repo)?;
-    {
-        let mut segments = url.path_segments_mut().map_err(|_| {
-            FirstPartyCapabilityError::new(RuntimeDispatchErrorKind::OperationFailed)
-        })?;
-        segments.push("git");
-        segments.push("matching-refs");
-        segments.push(namespace);
-        segments.push(prefix);
-    }
-    Ok(url)
-}
-
-fn github_api_headers() -> Vec<(String, String)> {
-    vec![
-        (
-            "User-Agent".to_string(),
-            "ironclaw-skill-install".to_string(),
-        ),
-        (
-            "Accept".to_string(),
-            "application/vnd.github+json".to_string(),
-        ),
-    ]
-}
-
-async fn fetch_github_api_value(
-    request: &FirstPartyCapabilityRequest,
-    url: &url::Url,
-    usage: &mut ResourceUsage,
-) -> Result<Value, FirstPartyCapabilityError> {
-    let bytes = fetch_url_bytes_with_headers(request, url, usage, github_api_headers()).await?;
-    serde_json::from_slice(&bytes).map_err(|_| {
-        FirstPartyCapabilityError::new(RuntimeDispatchErrorKind::OperationFailed)
-            .with_usage(usage.clone())
-    })
-}
-
-async fn resolve_github_default_branch(
-    request: &FirstPartyCapabilityRequest,
-    owner: &str,
-    repo: &str,
-    usage: &mut ResourceUsage,
-) -> Result<String, FirstPartyCapabilityError> {
-    let api_url = build_github_api_base_url(owner, repo)?;
-    let meta = fetch_github_api_value(request, &api_url, usage).await?;
-    meta.get("default_branch")
-        .and_then(Value::as_str)
-        .filter(|value| !value.trim().is_empty())
-        .map(str::to_string)
-        .ok_or_else(|| FirstPartyCapabilityError::new(RuntimeDispatchErrorKind::OperationFailed))
-}
-
-async fn resolve_github_ref_commit_sha(
-    request: &FirstPartyCapabilityRequest,
-    owner: &str,
-    repo: &str,
-    git_ref: &str,
-    usage: &mut ResourceUsage,
-) -> Result<String, FirstPartyCapabilityError> {
-    let mut commits_url = build_github_api_base_url(owner, repo)?;
-    {
-        let mut segments = commits_url.path_segments_mut().map_err(|_| {
-            FirstPartyCapabilityError::new(RuntimeDispatchErrorKind::OperationFailed)
-        })?;
-        segments.push("commits");
-    }
-    commits_url
-        .query_pairs_mut()
-        .append_pair("sha", git_ref)
-        .append_pair("per_page", "1");
-
-    let commits = fetch_github_api_value(request, &commits_url, usage).await?;
-    let sha = commits
-        .as_array()
-        .and_then(|commits| commits.first())
-        .and_then(|commit| commit.get("sha"))
-        .and_then(Value::as_str)
-        .filter(|sha| !sha.trim().is_empty())
-        .ok_or_else(|| FirstPartyCapabilityError::new(RuntimeDispatchErrorKind::OperationFailed))?;
-    if !sha.bytes().all(|b| b.is_ascii_hexdigit()) {
-        return Err(FirstPartyCapabilityError::new(
-            RuntimeDispatchErrorKind::OperationFailed,
-        ));
-    }
-    Ok(sha.to_string())
-}
-
-async fn resolve_github_ref_from_segments(
-    request: &FirstPartyCapabilityRequest,
-    owner: &str,
-    repo: &str,
-    segments: &[String],
-    usage: &mut ResourceUsage,
-) -> Result<(String, Vec<String>), FirstPartyCapabilityError> {
-    if segments.is_empty() || segments.len() > MAX_GITHUB_PATH_SEGMENTS {
-        return Err(FirstPartyCapabilityError::new(
-            RuntimeDispatchErrorKind::InputEncode,
-        ));
-    }
-    for namespace in ["heads", "tags"] {
-        let mut candidates = Vec::new();
-        let refs_url = build_github_matching_refs_url(owner, repo, namespace, &segments[0])?;
-        let refs = fetch_github_api_value(request, &refs_url, usage).await?;
-        let refs = refs.as_array().ok_or_else(|| {
-            FirstPartyCapabilityError::new(RuntimeDispatchErrorKind::OperationFailed)
-        })?;
-        let prefix = format!("refs/{namespace}/");
-        for item in refs {
-            let Some(git_ref) = item
-                .get("ref")
-                .and_then(Value::as_str)
-                .and_then(|value| value.strip_prefix(&prefix))
-            else {
-                continue;
-            };
-            let ref_segments = git_ref.split('/').collect::<Vec<_>>();
-            if ref_segments.len() > segments.len() {
-                continue;
-            }
-            if ref_segments
-                .iter()
-                .zip(segments.iter())
-                .all(|(left, right)| *left == right)
-            {
-                candidates.push((git_ref.to_string(), ref_segments.len()));
-            }
-        }
-        if let Some((git_ref, consumed)) =
-            candidates.into_iter().max_by_key(|(_, consumed)| *consumed)
-        {
-            return Ok((git_ref, segments[consumed..].to_vec()));
-        }
-    }
-    Err(FirstPartyCapabilityError::new(
-        RuntimeDispatchErrorKind::OperationFailed,
-    ))
-}
-
-async fn resolve_github_tree_request(
-    request: &FirstPartyCapabilityRequest,
-    repo: GitHubRepoRequest,
-    usage: &mut ResourceUsage,
-) -> Result<GitHubRepoRef, FirstPartyCapabilityError> {
-    validate_github_repo_components(&repo.owner, &repo.repo)?;
-    let branch = match repo.tree_segments {
-        Some(segments) => {
-            let (candidate_ref, remaining) = resolve_github_ref_from_segments(
-                request,
-                &repo.owner,
-                &repo.repo,
-                &segments,
-                usage,
-            )
-            .await?;
-            let candidate_subdir = (!remaining.is_empty()).then(|| remaining.join("/"));
-            return Ok(GitHubRepoRef {
-                owner: repo.owner,
-                repo: repo.repo,
-                branch: candidate_ref,
-                subdir: candidate_subdir,
-            });
-        }
-        None => resolve_github_default_branch(request, &repo.owner, &repo.repo, usage).await?,
-    };
-
-    Ok(GitHubRepoRef {
-        owner: repo.owner,
-        repo: repo.repo,
-        branch,
-        subdir: None,
-    })
-}
-
-async fn resolve_github_blob_download_url(
-    request: &FirstPartyCapabilityRequest,
-    blob: GitHubBlobRequest,
-    usage: &mut ResourceUsage,
-) -> Result<url::Url, FirstPartyCapabilityError> {
-    validate_github_repo_components(&blob.owner, &blob.repo)?;
-    if blob.blob_segments.len() < 2 || blob.blob_segments.len() > MAX_GITHUB_PATH_SEGMENTS {
-        return Err(FirstPartyCapabilityError::new(
-            RuntimeDispatchErrorKind::InputEncode,
-        ));
-    }
-    let (candidate_ref, remaining) = resolve_github_ref_from_segments(
-        request,
-        &blob.owner,
-        &blob.repo,
-        &blob.blob_segments,
-        usage,
-    )
-    .await?;
-    if remaining.is_empty() {
-        return Err(FirstPartyCapabilityError::new(
-            RuntimeDispatchErrorKind::InputEncode,
-        ));
-    }
-    let candidate_path = remaining.join("/");
-    let contents_url = build_github_contents_url(
-        &blob.owner,
-        &blob.repo,
-        Some(&candidate_path),
-        &candidate_ref,
-    )?;
-    let metadata = fetch_github_api_value(request, &contents_url, usage).await?;
-    if metadata.get("type").and_then(Value::as_str) != Some("file") {
-        return Err(FirstPartyCapabilityError::new(
-            RuntimeDispatchErrorKind::OperationFailed,
-        ));
-    }
-    let Some(download_url) = metadata.get("download_url").and_then(Value::as_str) else {
-        return Err(FirstPartyCapabilityError::new(
-            RuntimeDispatchErrorKind::OperationFailed,
-        ));
-    };
-    validate_derived_fetch_url(download_url)
-}
-
-async fn fetch_github_repo_payload(
-    request: &FirstPartyCapabilityRequest,
-    source_url: &str,
-    repo_request: GitHubRepoRequest,
-    usage: &mut ResourceUsage,
-) -> Result<SkillUrlPayload, FirstPartyCapabilityError> {
-    let repo = resolve_github_tree_request(request, repo_request, usage).await?;
-    if repo.subdir.is_some() {
-        return fetch_github_contents_bundle_payload(request, source_url, repo, usage).await;
-    }
-    let commit_sha =
-        resolve_github_ref_commit_sha(request, &repo.owner, &repo.repo, &repo.branch, usage)
-            .await?;
-    let archive_url = validate_derived_fetch_url(&format!(
-        "https://codeload.github.com/{}/{}/legacy.zip/{}",
-        repo.owner, repo.repo, commit_sha
-    ))?;
-    let bytes = fetch_url_bytes(request, &archive_url, usage).await?;
-    let bundle = extract_skill_bundle_from_zip_blocking(bytes, repo.subdir).await?;
-    tracing::debug!(
-        source_url,
-        source_subdir = ?bundle.bundle_subdir,
-        bundle_file_count = bundle.files.len(),
-        "skill URL install fetched GitHub bundle"
-    );
-    Ok(SkillUrlPayload {
-        content: bundle.skill_md,
-        files: bundle.files,
-    })
-}
-
-async fn fetch_github_contents_bundle_payload(
-    request: &FirstPartyCapabilityRequest,
-    source_url: &str,
-    repo: GitHubRepoRef,
-    usage: &mut ResourceUsage,
-) -> Result<SkillUrlPayload, FirstPartyCapabilityError> {
-    let root_subdir = repo
-        .subdir
-        .as_deref()
-        .ok_or_else(|| FirstPartyCapabilityError::new(RuntimeDispatchErrorKind::OperationFailed))?;
-    let root_path = normalize_archive_path(Path::new(root_subdir))?;
-    let mut directories = VecDeque::from([root_subdir.to_string()]);
-    let mut visited_directories = 0usize;
-    let mut seen_files = HashSet::<PathBuf>::new();
-    let mut skill_md = None;
-    let mut files = Vec::new();
-    let mut total_bytes = 0u64;
-
-    while let Some(directory) = directories.pop_front() {
-        visited_directories += 1;
-        if visited_directories > MAX_GITHUB_CONTENT_DIRS {
-            return Err(FirstPartyCapabilityError::new(
-                RuntimeDispatchErrorKind::OutputTooLarge,
-            ));
-        }
-        let contents_url =
-            build_github_contents_url(&repo.owner, &repo.repo, Some(&directory), &repo.branch)?;
-        let value = fetch_github_api_value(request, &contents_url, usage).await?;
-        let entries = value.as_array().ok_or_else(|| {
-            FirstPartyCapabilityError::new(RuntimeDispatchErrorKind::OperationFailed)
-        })?;
-        for entry in entries {
-            let entry_type = entry.get("type").and_then(Value::as_str);
-            let entry_path = entry.get("path").and_then(Value::as_str).ok_or_else(|| {
-                FirstPartyCapabilityError::new(RuntimeDispatchErrorKind::OperationFailed)
-            })?;
-            match entry_type {
-                Some("dir") => directories.push_back(entry_path.to_string()),
-                Some("file") => {
-                    let path = normalize_archive_path(Path::new(entry_path))?;
-                    let relative = path.strip_prefix(&root_path).map_err(|_| {
-                        FirstPartyCapabilityError::new(RuntimeDispatchErrorKind::OperationFailed)
-                    })?;
-                    if relative.as_os_str().is_empty() {
-                        continue;
-                    }
-                    if !seen_files.insert(relative.to_path_buf()) {
-                        return Err(FirstPartyCapabilityError::new(
-                            RuntimeDispatchErrorKind::InputEncode,
-                        ));
-                    }
-                    let download_url = entry
-                        .get("download_url")
-                        .and_then(Value::as_str)
-                        .ok_or_else(|| {
-                            FirstPartyCapabilityError::new(
-                                RuntimeDispatchErrorKind::OperationFailed,
-                            )
-                        })?;
-                    let download_url = validate_derived_fetch_url(download_url)?;
-                    let bytes = fetch_url_bytes(request, &download_url, usage).await?;
-                    if bytes.len() as u64 > MAX_ZIP_ENTRY_BYTES {
-                        return Err(FirstPartyCapabilityError::new(
-                            RuntimeDispatchErrorKind::OutputTooLarge,
-                        ));
-                    }
-                    total_bytes = total_bytes.checked_add(bytes.len() as u64).ok_or_else(|| {
-                        FirstPartyCapabilityError::new(RuntimeDispatchErrorKind::OutputTooLarge)
-                    })?;
-                    if total_bytes > MAX_TOTAL_UNZIPPED_BYTES {
-                        return Err(FirstPartyCapabilityError::new(
-                            RuntimeDispatchErrorKind::OutputTooLarge,
-                        ));
-                    }
-                    if relative == Path::new("SKILL.md") {
-                        if bytes.len() as u64 > ironclaw_skills::MAX_PROMPT_FILE_SIZE {
-                            return Err(FirstPartyCapabilityError::new(
-                                RuntimeDispatchErrorKind::OutputTooLarge,
-                            ));
-                        }
-                        skill_md = Some(String::from_utf8(bytes).map_err(|_| {
-                            FirstPartyCapabilityError::new(
-                                RuntimeDispatchErrorKind::OperationFailed,
-                            )
-                        })?);
-                    } else {
-                        files.push(SkillUrlPayloadFile {
-                            path: relative.to_path_buf(),
-                            contents: bytes,
-                        });
-                        if files.len() > ironclaw_skills::MAX_INSTALL_BUNDLE_FILES {
-                            return Err(FirstPartyCapabilityError::new(
-                                RuntimeDispatchErrorKind::OutputTooLarge,
-                            ));
-                        }
-                    }
-                }
-                _ => {}
-            }
-        }
-    }
-
-    let content = skill_md
-        .ok_or_else(|| FirstPartyCapabilityError::new(RuntimeDispatchErrorKind::OperationFailed))?;
-    tracing::debug!(
-        source_url,
-        source_subdir = root_subdir,
-        bundle_file_count = files.len(),
-        "skill URL install fetched GitHub contents bundle"
-    );
-    Ok(SkillUrlPayload { content, files })
-}
-
-async fn extract_skill_bundle_from_zip_blocking(
-    data: Vec<u8>,
-    requested_subdir: Option<String>,
-) -> Result<ZipSkillBundle, FirstPartyCapabilityError> {
-    tokio::task::spawn_blocking(move || {
-        extract_skill_bundle_from_zip(&data, requested_subdir.as_deref())
-    })
-    .await
-    .map_err(|error| {
-        if error.is_panic() {
-            tracing::error!("skill URL ZIP extraction worker panicked");
-        }
-        FirstPartyCapabilityError::new(RuntimeDispatchErrorKind::Backend)
-    })?
-}
-
-fn normalize_archive_path(path: &Path) -> Result<PathBuf, FirstPartyCapabilityError> {
-    let mut normalized = PathBuf::new();
-    for component in path.components() {
-        match component {
-            Component::Normal(part) => normalized.push(part),
-            Component::CurDir => {}
-            Component::ParentDir | Component::RootDir | Component::Prefix(_) => {
-                return Err(FirstPartyCapabilityError::new(
-                    RuntimeDispatchErrorKind::InputEncode,
-                ));
-            }
-        }
-    }
-    if normalized.as_os_str().is_empty() {
-        return Err(FirstPartyCapabilityError::new(
-            RuntimeDispatchErrorKind::InputEncode,
-        ));
-    }
-    Ok(normalized)
-}
-
-fn strip_common_archive_root(paths: &[PathBuf]) -> Option<PathBuf> {
-    let mut root: Option<std::ffi::OsString> = None;
-    let mut has_nested = false;
-    for path in paths {
-        let mut components = path.components();
-        let Some(Component::Normal(first)) = components.next() else {
-            return None;
-        };
-        has_nested |= components.next().is_some();
-        match &root {
-            Some(existing) if existing != first => return None,
-            None => root = Some(first.to_os_string()),
-            _ => {}
-        }
-    }
-    if !has_nested {
-        return None;
-    }
-    root.map(PathBuf::from)
-}
-
-fn extract_skill_bundle_from_zip(
-    data: &[u8],
-    requested_subdir: Option<&str>,
-) -> Result<ZipSkillBundle, FirstPartyCapabilityError> {
-    let reader = std::io::Cursor::new(data);
-    let mut archive = zip::ZipArchive::new(reader)
-        .map_err(|_| FirstPartyCapabilityError::new(RuntimeDispatchErrorKind::OperationFailed))?;
-
-    let mut raw_paths = Vec::new();
-    for index in 0..archive.len() {
-        let file = archive.by_index(index).map_err(|_| {
-            FirstPartyCapabilityError::new(RuntimeDispatchErrorKind::OperationFailed)
-        })?;
-        if !file.is_dir() {
-            if raw_paths.len() >= MAX_ZIP_FILE_ENTRIES {
-                return Err(FirstPartyCapabilityError::new(
-                    RuntimeDispatchErrorKind::OutputTooLarge,
-                ));
-            }
-            raw_paths.push(normalize_archive_path(Path::new(file.name()))?);
-        }
-    }
-    let strip_root = strip_common_archive_root(&raw_paths);
-    let mut files = Vec::<(PathBuf, Vec<u8>)>::new();
-    let mut seen_paths = std::collections::HashSet::<PathBuf>::new();
-    let mut skill_dirs = std::collections::HashSet::<PathBuf>::new();
-    let mut total_unzipped_bytes = 0u64;
-
-    for index in 0..archive.len() {
-        let mut file = archive.by_index(index).map_err(|_| {
-            FirstPartyCapabilityError::new(RuntimeDispatchErrorKind::OperationFailed)
-        })?;
-        if file.is_dir() {
-            continue;
-        }
-        if file.size() > MAX_ZIP_ENTRY_BYTES {
-            return Err(FirstPartyCapabilityError::new(
-                RuntimeDispatchErrorKind::OutputTooLarge,
-            ));
-        }
-        let entry_name = file.name().to_string();
-        let mut path = normalize_archive_path(Path::new(&entry_name))?;
-        if let Some(root) = &strip_root
-            && let Ok(stripped) = path.strip_prefix(root)
-        {
-            path = stripped.to_path_buf();
-        }
-        if path.as_os_str().is_empty() {
-            continue;
-        }
-        if !seen_paths.insert(path.clone()) {
-            return Err(FirstPartyCapabilityError::new(
-                RuntimeDispatchErrorKind::InputEncode,
-            ));
-        }
-
-        let mut contents = Vec::new();
-        (&mut file)
-            .take(MAX_ZIP_ENTRY_BYTES + 1)
-            .read_to_end(&mut contents)
-            .map_err(|_| {
-                FirstPartyCapabilityError::new(RuntimeDispatchErrorKind::OperationFailed)
-            })?;
-        if contents.len() as u64 > MAX_ZIP_ENTRY_BYTES {
-            return Err(FirstPartyCapabilityError::new(
-                RuntimeDispatchErrorKind::OutputTooLarge,
-            ));
-        }
-        total_unzipped_bytes = total_unzipped_bytes
-            .checked_add(contents.len() as u64)
-            .ok_or_else(|| {
-                FirstPartyCapabilityError::new(RuntimeDispatchErrorKind::OutputTooLarge)
-            })?;
-        if total_unzipped_bytes > MAX_TOTAL_UNZIPPED_BYTES {
-            return Err(FirstPartyCapabilityError::new(
-                RuntimeDispatchErrorKind::OutputTooLarge,
-            ));
-        }
-        if path.file_name().is_some_and(|name| name == "SKILL.md") {
-            skill_dirs.insert(path.parent().unwrap_or(Path::new("")).to_path_buf());
-        }
-        files.push((path, contents));
-    }
-
-    let requested_dir = if let Some(subdir) = requested_subdir {
-        let normalized = normalize_archive_path(Path::new(subdir))?;
-        if !skill_dirs.contains(&normalized) {
-            return Err(FirstPartyCapabilityError::new(
-                RuntimeDispatchErrorKind::OperationFailed,
-            ));
-        }
-        normalized
-    } else {
-        match skill_dirs.len() {
-            0 => {
-                return Err(FirstPartyCapabilityError::new(
-                    RuntimeDispatchErrorKind::OperationFailed,
-                ));
-            }
-            1 => skill_dirs.into_iter().next().unwrap_or_default(),
-            _ => {
-                return Err(FirstPartyCapabilityError::new(
-                    RuntimeDispatchErrorKind::InputEncode,
-                ));
-            }
-        }
-    };
-
-    let mut skill_md = None;
-    let mut extra_files = Vec::new();
-    for (path, contents) in files {
-        let Ok(relative) = path.strip_prefix(&requested_dir) else {
-            continue;
-        };
-        if relative.as_os_str().is_empty() {
-            continue;
-        }
-        if relative == Path::new("SKILL.md") {
-            if contents.len() as u64 > ironclaw_skills::MAX_PROMPT_FILE_SIZE {
-                return Err(FirstPartyCapabilityError::new(
-                    RuntimeDispatchErrorKind::OutputTooLarge,
-                ));
-            }
-            skill_md = Some(String::from_utf8(contents).map_err(|_| {
-                FirstPartyCapabilityError::new(RuntimeDispatchErrorKind::OperationFailed)
-            })?);
-            continue;
-        }
-        extra_files.push(SkillUrlPayloadFile {
-            path: relative.to_path_buf(),
-            contents,
-        });
-        if extra_files.len() > ironclaw_skills::MAX_INSTALL_BUNDLE_FILES {
-            return Err(FirstPartyCapabilityError::new(
-                RuntimeDispatchErrorKind::OutputTooLarge,
-            ));
-        }
-    }
-
-    let skill_md = skill_md
-        .ok_or_else(|| FirstPartyCapabilityError::new(RuntimeDispatchErrorKind::OperationFailed))?;
-    Ok(ZipSkillBundle {
-        skill_md,
-        files: extra_files,
-        bundle_subdir: (!requested_dir.as_os_str().is_empty())
-            .then(|| requested_dir.display().to_string()),
-    })
 }
 
 fn skill_url_fetch_error(
