+++ crates/ironclaw_skills/src/management.rs
@@ -1,22 +1,27 @@
-use std::{path::Component, sync::Arc};
+use std::sync::Arc;
 
 use crate::{MAX_PROMPT_FILE_SIZE, normalize_line_endings, parse_skill_md, validate_skill_name};
 use async_trait::async_trait;
 use ironclaw_filesystem::{
-    BackendCapabilities, DirEntry, FileStat, FileType, FilesystemError, FilesystemOperation,
-    RootFilesystem, ScopedFilesystem,
+    BackendCapabilities, DirEntry, FileStat, FileType, FilesystemError, RootFilesystem,
+    ScopedFilesystem,
 };
 use ironclaw_host_api::{HostApiError, MountView, ResourceScope, ScopedPath, VirtualPath};
-use serde::Deserialize;
-use serde_json::json;
 
-const USER_SKILLS_ROOT: &str = "/skills";
+mod install_bundle;
+#[cfg(test)]
+mod tests;
+
+pub use install_bundle::{MAX_INSTALL_BUNDLE_FILES, SkillInstallFile};
+
+use install_bundle::{
+    install_metadata_source, installed_skill_source, publish_skill_install,
+    read_install_metadata_bytes, validate_install_bundle_files,
+};
+
+pub(super) const USER_SKILLS_ROOT: &str = "/skills";
 const SYSTEM_SKILLS_ROOT: &str = "/system/skills";
-const SKILL_FILE_NAME: &str = "SKILL.md";
-const INSTALL_METADATA_FILE_NAME: &str = ".ironclaw-install.json";
-pub const MAX_INSTALL_BUNDLE_FILES: usize = 256;
-const MAX_INSTALL_BUNDLE_FILE_BYTES: usize = 2 * 1024 * 1024;
-const MAX_INSTALL_BUNDLE_TOTAL_BYTES: usize = 20 * 1024 * 1024;
+pub(super) const SKILL_FILE_NAME: &str = "SKILL.md";
 
 #[derive(Debug, Clone, Copy, PartialEq, Eq)]
 pub enum SkillManagementErrorKind {
@@ -148,12 +153,6 @@ pub struct SkillInstallRequest<'a> {
     pub source_url: Option<&'a str>,
 }
 
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-pub struct SkillInstallFile<'a> {
-    pub relative_path: &'a str,
-    pub contents: &'a [u8],
-}
-
 #[derive(Debug, Clone, PartialEq, Eq)]
 pub struct SkillInstallResult {
     pub name: String,
@@ -225,7 +224,6 @@ pub async fn install_skill(
     validate_install_bundle_files(request.files)?;
 
     let skill_name = parsed.manifest.name;
-    let skill_dir = skill_root_scoped_path(USER_SKILLS_ROOT, &skill_name)?;
     let skill_path = skill_scoped_path(USER_SKILLS_ROOT, &skill_name, SKILL_FILE_NAME)?;
 
     log_skill_filesystem_phase("stat_existing", &skill_name, &skill_path);
@@ -240,94 +238,15 @@ pub async fn install_skill(
         ));
     }
 
-    log_skill_filesystem_phase("create_dir_all", &skill_name, &skill_dir);
-    context
-        .filesystem
-        .create_dir_all(&context.scope, &skill_dir)
-        .await
-        .or_else(|error| match error {
-            FilesystemError::Unsupported {
-                operation: FilesystemOperation::CreateDirAll,
-                ..
-            } => {
-                log_skill_filesystem_phase("create_dir_all_unsupported", &skill_name, &skill_dir);
-                Ok(())
-            }
-            other => Err(other),
-        })
-        .map_err(|error| {
-            log_skill_filesystem_phase("create_dir_all_failed", &skill_name, &skill_dir);
-            filesystem_error(error)
-        })?;
-    for file in request.files {
-        let relative_path = normalize_install_relative_path(file.relative_path)?;
-        let file_path =
-            skill_bundle_file_scoped_path(USER_SKILLS_ROOT, &skill_name, &relative_path)?;
-        if let Some(parent) = scoped_parent(&file_path)? {
-            log_skill_filesystem_phase("create_bundle_parent", &skill_name, &parent);
-            context
-                .filesystem
-                .create_dir_all(&context.scope, &parent)
-                .await
-                .or_else(|error| match error {
-                    FilesystemError::Unsupported {
-                        operation: FilesystemOperation::CreateDirAll,
-                        ..
-                    } => {
-                        log_skill_filesystem_phase(
-                            "create_bundle_parent_unsupported",
-                            &skill_name,
-                            &parent,
-                        );
-                        Ok(())
-                    }
-                    other => Err(other),
-                })
-                .map_err(|error| {
-                    log_skill_filesystem_phase("create_bundle_parent_failed", &skill_name, &parent);
-                    filesystem_error(error)
-                })?;
-        }
-        log_skill_filesystem_phase("write_bundle_file", &skill_name, &file_path);
-        context
-            .filesystem
-            .write_file(&context.scope, &file_path, file.contents)
-            .await
-            .map_err(|error| {
-                log_skill_filesystem_phase("write_bundle_file_failed", &skill_name, &file_path);
-                filesystem_error(error)
-            })?;
-    }
-    if request.source == SkillInstallSource::InstalledUrl {
-        let metadata_path = skill_bundle_file_scoped_path(
-            USER_SKILLS_ROOT,
-            &skill_name,
-            INSTALL_METADATA_FILE_NAME,
-        )?;
-        let metadata = install_metadata_bytes(request.source_url)?;
-        log_skill_filesystem_phase("write_install_metadata", &skill_name, &metadata_path);
-        context
-            .filesystem
-            .write_file(&context.scope, &metadata_path, &metadata)
-            .await
-            .map_err(|error| {
-                log_skill_filesystem_phase(
-                    "write_install_metadata_failed",
-                    &skill_name,
-                    &metadata_path,
-                );
-                filesystem_error(error)
-            })?;
-    }
-    log_skill_filesystem_phase("write_file", &skill_name, &skill_path);
-    context
-        .filesystem
-        .write_file(&context.scope, &skill_path, normalized.as_bytes())
-        .await
-        .map_err(|error| {
-            log_skill_filesystem_phase("write_file_failed", &skill_name, &skill_path);
-            filesystem_error(error)
-        })?;
+    publish_skill_install(
+        context,
+        &skill_name,
+        &normalized,
+        request.files,
+        request.source,
+        request.source_url,
+    )
+    .await?;
     tracing::debug!(
         skill_name = %skill_name,
         scoped_path = %skill_path,
@@ -468,11 +387,6 @@ async fn read_skill_summary(
     }))
 }
 
-#[derive(Debug, Deserialize)]
-struct InstallMetadata {
-    source: Option<String>,
-}
-
 fn skill_root_scoped_path(root: &str, name: &str) -> Result<ScopedPath, SkillManagementError> {
     skill_scoped_path(root, name, "")
 }
@@ -504,112 +418,6 @@ fn skill_scoped_path(
         .map_err(|_| SkillManagementError::new(SkillManagementErrorKind::InvalidInput))
 }
 
-fn skill_bundle_file_scoped_path(
-    root: &str,
-    name: &str,
-    relative_path: &str,
-) -> Result<ScopedPath, SkillManagementError> {
-    if !validate_skill_name(name) {
-        return Err(SkillManagementError::new(
-            SkillManagementErrorKind::InvalidInput,
-        ));
-    }
-    ScopedPath::new(format!(
-        "{}/{}/{}",
-        root.trim_end_matches('/'),
-        name,
-        relative_path
-    ))
-    .map_err(|_| SkillManagementError::new(SkillManagementErrorKind::InvalidInput))
-}
-
-fn validate_install_bundle_files(
-    files: &[SkillInstallFile<'_>],
-) -> Result<(), SkillManagementError> {
-    if files.len() > MAX_INSTALL_BUNDLE_FILES {
-        return Err(SkillManagementError::new(
-            SkillManagementErrorKind::Resource,
-        ));
-    }
-    let mut total_bytes = 0usize;
-    for file in files {
-        if file.contents.len() > MAX_INSTALL_BUNDLE_FILE_BYTES {
-            return Err(SkillManagementError::new(
-                SkillManagementErrorKind::Resource,
-            ));
-        }
-        total_bytes = total_bytes
-            .checked_add(file.contents.len())
-            .ok_or_else(|| SkillManagementError::new(SkillManagementErrorKind::Resource))?;
-        if total_bytes > MAX_INSTALL_BUNDLE_TOTAL_BYTES {
-            return Err(SkillManagementError::new(
-                SkillManagementErrorKind::Resource,
-            ));
-        }
-        normalize_install_relative_path(file.relative_path)?;
-    }
-    Ok(())
-}
-
-fn normalize_install_relative_path(path: &str) -> Result<String, SkillManagementError> {
-    if path.is_empty()
-        || path.starts_with('/')
-        || path.contains('\\')
-        || path.contains('\0')
-        || path.chars().any(char::is_control)
-        || path.contains("://")
-    {
-        return Err(SkillManagementError::new(
-            SkillManagementErrorKind::InvalidInput,
-        ));
-    }
-
-    let mut parts = Vec::new();
-    for component in std::path::Path::new(path).components() {
-        match component {
-            Component::Normal(part) => {
-                let part = part.to_str().ok_or_else(|| {
-                    SkillManagementError::new(SkillManagementErrorKind::InvalidInput)
-                })?;
-                if part.is_empty() {
-                    return Err(SkillManagementError::new(
-                        SkillManagementErrorKind::InvalidInput,
-                    ));
-                }
-                parts.push(part);
-            }
-            Component::CurDir => {}
-            Component::ParentDir | Component::RootDir | Component::Prefix(_) => {
-                return Err(SkillManagementError::new(
-                    SkillManagementErrorKind::InvalidInput,
-                ));
-            }
-        }
-    }
-
-    if parts.is_empty() || parts == [SKILL_FILE_NAME] || parts == [INSTALL_METADATA_FILE_NAME] {
-        return Err(SkillManagementError::new(
-            SkillManagementErrorKind::InvalidInput,
-        ));
-    }
-    Ok(parts.join("/"))
-}
-
-fn installed_skill_source(source: SkillInstallSource) -> SkillSource {
-    match source {
-        SkillInstallSource::User => SkillSource::User,
-        SkillInstallSource::InstalledUrl => SkillSource::Installed,
-    }
-}
-
-fn install_metadata_bytes(source_url: Option<&str>) -> Result<Vec<u8>, SkillManagementError> {
-    serde_json::to_vec_pretty(&json!({
-        "source": "installed_url",
-        "source_url": source_url,
-    }))
-    .map_err(|_| SkillManagementError::new(SkillManagementErrorKind::InvalidInput))
-}
-
 async fn skill_source_with_install_metadata(
     context: &SkillManagementContext,
     skill_path: &ScopedPath,
@@ -618,25 +426,10 @@ async fn skill_source_with_install_metadata(
     if default_source != SkillSource::User {
         return Ok(default_source);
     }
-    let Some(metadata_path) = scoped_sibling(skill_path, INSTALL_METADATA_FILE_NAME)? else {
+    let Some(bytes) = read_install_metadata_bytes(context, skill_path).await? else {
         return Ok(default_source);
     };
-    let bytes = match context
-        .filesystem
-        .read_bytes_bounded(&context.scope, &metadata_path, 4096)
-        .await
-    {
-        Ok(Some(bytes)) => bytes,
-        Ok(None) | Err(FilesystemError::NotFound { .. }) => return Ok(default_source),
-        Err(error) => return Err(filesystem_error(error)),
-    };
-    let Ok(metadata) = serde_json::from_slice::<InstallMetadata>(&bytes) else {
-        return Ok(SkillSource::Installed);
-    };
-    match metadata.source.as_deref() {
-        Some("installed_url") => Ok(SkillSource::Installed),
-        Some(_) | None => Ok(SkillSource::Installed),
-    }
+    Ok(install_metadata_source(default_source, &bytes))
 }
 
 fn scoped_sibling(
@@ -650,18 +443,6 @@ fn scoped_sibling(
         return Ok(None);
     }
     ScopedPath::new(format!("{parent}/{sibling}"))
-        .map(Some)
-        .map_err(|_| SkillManagementError::new(SkillManagementErrorKind::InvalidInput))
-}
-
-fn scoped_parent(path: &ScopedPath) -> Result<Option<ScopedPath>, SkillManagementError> {
-    let Some((parent, _)) = path.as_str().rsplit_once('/') else {
-        return Ok(None);
-    };
-    if parent.is_empty() || parent == USER_SKILLS_ROOT {
-        return Ok(None);
-    }
-    ScopedPath::new(parent.to_string())
         .map(Some)
         .map_err(|_| SkillManagementError::new(SkillManagementErrorKind::InvalidInput))
 }
@@ -755,438 +536,5 @@ fn filesystem_error(error: FilesystemError) -> SkillManagementError {
             SkillManagementError::new(SkillManagementErrorKind::FilesystemDenied)
         }
         _ => SkillManagementError::new(SkillManagementErrorKind::FilesystemDenied),
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use ironclaw_filesystem::InMemoryBackend;
-    use ironclaw_host_api::{MountAlias, MountGrant, MountPermissions};
-
-    #[tokio::test]
-    async fn install_list_and_remove_user_skills_through_scoped_mounts() {
-        let filesystem = Arc::new(InMemoryBackend::default());
-        write_file(
-            filesystem.as_ref(),
-            "/projects/system/skills/system-helper/SKILL.md",
-            skill_md(
-                "system-helper",
-                "system skill description",
-                "SYSTEM_SKILL_PROMPT",
-            ),
-        )
-        .await;
-        let context = skill_management_context(filesystem.clone(), skill_mounts());
-
-        let installed = install_skill(
-            &context,
-            SkillInstallRequest {
-                name: None,
-                content: &skill_md(
-                    "local-helper",
-                    "local skill description",
-                    "LOCAL_SKILL_PROMPT",
-                ),
-                files: &[],
-                source: SkillInstallSource::User,
-                source_url: None,
-            },
-        )
-        .await
-        .unwrap();
-        assert_eq!(installed.name, "local-helper");
-        assert_eq!(
-            installed.scoped_path,
-            "/skills/local-helper/SKILL.md".to_string()
-        );
-
-        let listed = list_skills(&context).await.unwrap();
-        assert_eq!(listed.len(), 2);
-        assert!(
-            listed
-                .iter()
-                .any(|skill| skill.name == "system-helper" && skill.source == SkillSource::System)
-        );
-        assert!(
-            listed
-                .iter()
-                .any(|skill| skill.name == "local-helper" && skill.source == SkillSource::User)
-        );
-
-        let removed = remove_skill(
-            &context,
-            SkillRemoveRequest {
-                name: "local-helper",
-            },
-        )
-        .await
-        .unwrap();
-        assert_eq!(removed.name, "local-helper");
-        assert_eq!(list_skills(&context).await.unwrap().len(), 1);
-    }
-
-    #[tokio::test]
-    async fn install_rejects_name_mismatch() {
-        let filesystem = Arc::new(InMemoryBackend::default());
-        let context = skill_management_context(filesystem, skill_mounts());
-
-        let error = install_skill(
-            &context,
-            SkillInstallRequest {
-                name: Some("expected"),
-                content: &skill_md("actual", "description", "PROMPT"),
-                files: &[],
-                source: SkillInstallSource::User,
-                source_url: None,
-            },
-        )
-        .await
-        .unwrap_err();
-
-        assert_eq!(error.kind(), SkillManagementErrorKind::InvalidInput);
-    }
-
-    #[tokio::test]
-    async fn install_rejects_invalid_bundle_files() {
-        let cases = [
-            (
-                "../escape.md",
-                b"ok".as_slice(),
-                SkillManagementErrorKind::InvalidInput,
-            ),
-            (
-                "/absolute.md",
-                b"ok".as_slice(),
-                SkillManagementErrorKind::InvalidInput,
-            ),
-            (
-                "SKILL.md",
-                b"ok".as_slice(),
-                SkillManagementErrorKind::InvalidInput,
-            ),
-            (
-                ".ironclaw-install.json",
-                b"ok".as_slice(),
-                SkillManagementErrorKind::InvalidInput,
-            ),
-        ];
-
-        for (relative_path, contents, expected) in cases {
-            let filesystem = Arc::new(InMemoryBackend::default());
-            let context = skill_management_context(filesystem, skill_mounts());
-
-            let error = install_skill(
-                &context,
-                SkillInstallRequest {
-                    name: None,
-                    content: &skill_md("bundle-helper", "description", "PROMPT"),
-                    files: &[SkillInstallFile {
-                        relative_path,
-                        contents,
-                    }],
-                    source: SkillInstallSource::User,
-                    source_url: None,
-                },
-            )
-            .await
-            .unwrap_err();
-
-            assert_eq!(error.kind(), expected);
-        }
-
-        let oversized = vec![b'x'; MAX_INSTALL_BUNDLE_FILE_BYTES + 1];
-        let filesystem = Arc::new(InMemoryBackend::default());
-        let context = skill_management_context(filesystem, skill_mounts());
-        let error = install_skill(
-            &context,
-            SkillInstallRequest {
-                name: None,
-                content: &skill_md("oversized-helper", "description", "PROMPT"),
-                files: &[SkillInstallFile {
-                    relative_path: "references/large.bin",
-                    contents: &oversized,
-                }],
-                source: SkillInstallSource::User,
-                source_url: None,
-            },
-        )
-        .await
-        .unwrap_err();
-        assert_eq!(error.kind(), SkillManagementErrorKind::Resource);
-
-        let paths = (0..=MAX_INSTALL_BUNDLE_FILES)
-            .map(|index| format!("references/{index}.md"))
-            .collect::<Vec<_>>();
-        let files = paths
-            .iter()
-            .map(|path| SkillInstallFile {
-                relative_path: path.as_str(),
-                contents: b"ok",
-            })
-            .collect::<Vec<_>>();
-        let filesystem = Arc::new(InMemoryBackend::default());
-        let context = skill_management_context(filesystem, skill_mounts());
-        let error = install_skill(
-            &context,
-            SkillInstallRequest {
-                name: None,
-                content: &skill_md("too-many-helper", "description", "PROMPT"),
-                files: &files,
-                source: SkillInstallSource::User,
-                source_url: None,
-            },
-        )
-        .await
-        .unwrap_err();
-        assert_eq!(error.kind(), SkillManagementErrorKind::Resource);
-    }
-
-    #[tokio::test]
-    async fn install_bundle_failure_does_not_publish_skill_md() {
-        let inner = Arc::new(InMemoryBackend::default());
-        let filesystem = Arc::new(FailingBundleWriteFilesystem {
-            inner: inner.clone(),
-            fail_suffix: "/scripts/run.py",
-        });
-        let context = skill_management_context_with_root(filesystem, skill_mounts());
-
-        let error = install_skill(
-            &context,
-            SkillInstallRequest {
-                name: None,
-                content: &skill_md("partial-helper", "description", "PROMPT"),
-                files: &[SkillInstallFile {
-                    relative_path: "scripts/run.py",
-                    contents: b"print('nope')\n",
-                }],
-                source: SkillInstallSource::User,
-                source_url: None,
-            },
-        )
-        .await
-        .unwrap_err();
-        assert_eq!(error.kind(), SkillManagementErrorKind::InvalidSkill);
-
-        match inner
-            .read_file_bounded(
-                &VirtualPath::new("/projects/skills/partial-helper/SKILL.md").unwrap(),
-                1024,
-            )
-            .await
-        {
-            Ok(None) | Err(FilesystemError::NotFound { .. }) => {}
-            Ok(Some(_)) => panic!(
-                "SKILL.md should be written last so failed bundle writes do not publish a partial skill"
-            ),
-            Err(error) => panic!("unexpected filesystem error: {error:?}"),
-        }
-    }
-
-    #[tokio::test]
-    async fn install_metadata_write_failure_does_not_publish_skill_md() {
-        let inner = Arc::new(InMemoryBackend::default());
-        let filesystem = Arc::new(FailingBundleWriteFilesystem {
-            inner: inner.clone(),
-            fail_suffix: "/.ironclaw-install.json",
-        });
-        let context = skill_management_context_with_root(filesystem, skill_mounts());
-
-        let error = install_skill(
-            &context,
-            SkillInstallRequest {
-                name: None,
-                content: &skill_md("metadata-helper", "description", "PROMPT"),
-                files: &[],
-                source: SkillInstallSource::InstalledUrl,
-                source_url: Some("https://example.test/SKILL.md"),
-            },
-        )
-        .await
-        .unwrap_err();
-        assert_eq!(error.kind(), SkillManagementErrorKind::InvalidSkill);
-
-        match inner
-            .read_file_bounded(
-                &VirtualPath::new("/projects/skills/metadata-helper/SKILL.md").unwrap(),
-                1024,
-            )
-            .await
-        {
-            Ok(None) | Err(FilesystemError::NotFound { .. }) => {}
-            Ok(Some(_)) => panic!(
-                "SKILL.md should be written last so failed metadata writes do not publish a partial skill"
-            ),
-            Err(error) => panic!("unexpected filesystem error: {error:?}"),
-        }
-    }
-
-    #[tokio::test]
-    async fn list_treats_malformed_install_metadata_as_installed() {
-        let filesystem = Arc::new(InMemoryBackend::default());
-        write_file(
-            filesystem.as_ref(),
-            "/projects/skills/metadata-helper/SKILL.md",
-            skill_md("metadata-helper", "local skill description", "PROMPT"),
-        )
-        .await;
-        filesystem
-            .write_file(
-                &VirtualPath::new("/projects/skills/metadata-helper/.ironclaw-install.json")
-                    .unwrap(),
-                b"not json",
-            )
-            .await
-            .unwrap();
-        let context = skill_management_context(filesystem, skill_mounts());
-
-        let listed = list_skills(&context).await.unwrap();
-
-        assert_eq!(listed.len(), 1);
-        assert_eq!(listed[0].name, "metadata-helper");
-        assert_eq!(listed[0].source, SkillSource::Installed);
-    }
-
-    #[tokio::test]
-    async fn list_treats_unmounted_optional_skill_root_as_empty() {
-        let filesystem = Arc::new(InMemoryBackend::default());
-        write_file(
-            filesystem.as_ref(),
-            "/projects/skills/local-helper/SKILL.md",
-            skill_md("local-helper", "local skill description", "PROMPT"),
-        )
-        .await;
-        let context = skill_management_context(filesystem, user_skill_mounts());
-
-        let listed = list_skills(&context).await.unwrap();
-
-        assert_eq!(listed.len(), 1);
-        assert_eq!(listed[0].name, "local-helper");
-        assert_eq!(listed[0].source, SkillSource::User);
-    }
-
-    #[tokio::test]
-    async fn remove_rejects_system_skill() {
-        let filesystem = Arc::new(InMemoryBackend::default());
-        write_file(
-            filesystem.as_ref(),
-            "/projects/system/skills/system-helper/SKILL.md",
-            skill_md("system-helper", "system skill description", "PROMPT"),
-        )
-        .await;
-        let context = skill_management_context(filesystem, skill_mounts());
-
-        let error = remove_skill(
-            &context,
-            SkillRemoveRequest {
-                name: "system-helper",
-            },
-        )
-        .await
-        .unwrap_err();
-
-        assert_eq!(error.kind(), SkillManagementErrorKind::NotFound);
-    }
-
-    async fn write_file(root: &InMemoryBackend, path: &str, body: String) {
-        root.write_file(&VirtualPath::new(path).unwrap(), body.as_bytes())
-            .await
-            .unwrap();
-    }
-
-    #[derive(Clone)]
-    struct FailingBundleWriteFilesystem {
-        inner: Arc<InMemoryBackend>,
-        fail_suffix: &'static str,
-    }
-
-    #[async_trait]
-    impl RootFilesystem for FailingBundleWriteFilesystem {
-        fn capabilities(&self) -> BackendCapabilities {
-            self.inner.capabilities()
-        }
-
-        async fn list_dir(&self, path: &VirtualPath) -> Result<Vec<DirEntry>, FilesystemError> {
-            self.inner.list_dir(path).await
-        }
-
-        async fn stat(&self, path: &VirtualPath) -> Result<FileStat, FilesystemError> {
-            self.inner.stat(path).await
-        }
-
-        async fn read_file_bounded(
-            &self,
-            path: &VirtualPath,
-            max_bytes: usize,
-        ) -> Result<Option<Vec<u8>>, FilesystemError> {
-            self.inner.read_file_bounded(path, max_bytes).await
-        }
-
-        async fn write_file(
-            &self,
-            path: &VirtualPath,
-            bytes: &[u8],
-        ) -> Result<(), FilesystemError> {
-            if path.as_str().ends_with(self.fail_suffix) {
-                return Err(FilesystemError::Backend {
-                    operation: FilesystemOperation::WriteFile,
-                    path: path.clone(),
-                    reason: "injected bundle write failure".to_string(),
-                });
-            }
-            self.inner.write_file(path, bytes).await
-        }
-
-        async fn create_dir_all(&self, path: &VirtualPath) -> Result<(), FilesystemError> {
-            self.inner.create_dir_all(path).await
-        }
-
-        async fn delete(&self, path: &VirtualPath) -> Result<(), FilesystemError> {
-            self.inner.delete(path).await
-        }
-    }
-
-    fn skill_mounts() -> MountView {
-        MountView::new(vec![
-            MountGrant::new(
-                MountAlias::new("/skills").unwrap(),
-                VirtualPath::new("/projects/skills").unwrap(),
-                MountPermissions::read_write_list_delete(),
-            ),
-            MountGrant::new(
-                MountAlias::new("/system/skills").unwrap(),
-                VirtualPath::new("/projects/system/skills").unwrap(),
-                MountPermissions::read_only(),
-            ),
-        ])
-        .unwrap()
-    }
-
-    fn user_skill_mounts() -> MountView {
-        MountView::new(vec![MountGrant::new(
-            MountAlias::new("/skills").unwrap(),
-            VirtualPath::new("/projects/skills").unwrap(),
-            MountPermissions::read_write_list_delete(),
-        )])
-        .unwrap()
-    }
-
-    fn skill_management_context(
-        filesystem: Arc<InMemoryBackend>,
-        mounts: MountView,
-    ) -> SkillManagementContext {
-        let filesystem: Arc<dyn RootFilesystem> = filesystem;
-        SkillManagementContext::new(filesystem, mounts, ResourceScope::system())
-    }
-
-    fn skill_management_context_with_root(
-        filesystem: Arc<dyn RootFilesystem>,
-        mounts: MountView,
-    ) -> SkillManagementContext {
-        SkillManagementContext::new(filesystem, mounts, ResourceScope::system())
-    }
-
-    fn skill_md(name: &str, description: &str, prompt: &str) -> String {
-        format!("---\nname: {name}\ndescription: {description}\n---\n{prompt}\n")
     }
 }
